p4i integration Integrate 255492 from 4.3 to main: RevBy:    Trust me Details:  On Windows we now also use the shared function for determining the writing system                  for a font.
[git-p4: depot-paths = "//depot/qt/main/": change = 255494]

+++ src/gui/text/qfontdatabase_qws.cpp
@@ -162,10 +162,10 @@ QStringList QFontDatabasePrivate::addTTFile(const QByteArray &file, const QByteA
         if (writingSystems.isEmpty()) {
             TT_OS2 *os2 = (TT_OS2 *)FT_Get_Sfnt_Table(face, ft_sfnt_os2);
             if (os2) {
-                ushort unicodeRange[4] = {
+                quint32 unicodeRange[4] = {
                     os2->ulUnicodeRange1, os2->ulUnicodeRange2, os2->ulUnicodeRange3, os2->ulUnicodeRange4
                 };
-                ushort codePageRange[2] = {
+                quint32 codePageRange[2] = {
                     os2->ulCodePageRange1, os2->ulCodePageRange2
                 };
